refactored main.promptForUpdateIfRequired to update.PromptFor
